docs(fix): better usability on feedback form (#10382)
### Description

I went back to use our feedback form to test it out and realized the
behavior and user feedback feels awkward. You can type a message, but
its not clear that you need to select an emoji before you can submit the
form.

This PR adds a small message to ensure the required action for form
submission is clear.

### Testing Instructions

Try out the feedaback form on the preview.