fix(microfrontends): respect packageName field (#10383)
### Description

This PR does 2 things:
- Adds a warning if a package is referenced in `microfrontends.json`
that is not present in the monorepo
 - Uses `packageName` over the `applications` key if present

### Testing Instructions

Added unit tests.

Verified warnings now are present if you reference a package that isn't
found in the monorepo:
```
[0 olszewski@macbookpro] /Users/olszewski/code/vercel/microfrontends $ turbo_dev --skip-infer dev -F nextjs-app-marketing -F nextjs-app-docs --dry=json > /dev/null
turbo 2.5.1

 WARNING  Unable to find packages referenced in 'microfrontends.json' in workspace.Local proxy will not route to the following applications if they are running locally: nextjs-app-marketing-test
```

Verified that proxy is now started if every application uses a
`packageName` that differs from the key in `applications`.

---------

Co-authored-by: Thomas Knickman <tom.knickman@vercel.com>