feat: add `Workspace.find_package_by_path` (#10493)
### Description

As raised
[here](https://linear.app/vercel/issue/TURBO-4724/expose-package-path-in-turborepository).

Depending on the call-pattern that we want to support, it might be
worthwhile to introduce a cache for ancestor-paths so that subsequent
calls only need to traverse-up to a first-common-ancestor.

### Testing Instructions

`pnpm build && pnpm test` within `packages/turbo-repository` passes.

### Note

I'm very much a Rust newbie, so please don't hold back on stylistic nits
or comments - I have no strong opinions on "correct" Rust!

---------

Co-authored-by: Chris Olszewski <chris.olszewski@vercel.com>