fix(pnpm): respect deep for linkWorkspacePackages (#10558)
### Description

Closes https://github.com/vercel/turborepo/issues/10528

With https://github.com/vercel/turborepo/pull/10391 we started reading
`linkWorkspacePackages` from `pnpm-workspace.yaml`. This caused issues
with single package vs monorepo detection as `deep` is a valid value
here and we only expected a boolean. This would cause a detection of a
single package repo and we would swallow the error of
`pnpm-workspace.yaml` not parsing due to the usage of `deep`.

This PR adds support for [`deep` on the parsing
end](https://pnpm.io/settings#linkworkspacepackages)

Follow up PR is to have noisy warnings if we find a
`pnpm-workspace.yaml`, but do not understand it.

### Testing Instructions

Added a red/green test. Verified with
https://github.com/ocavue/starter-monorepo