chore: add unit test coverage reporter (#10641)
### Description

I've been using LLMs to write code a lot lately and realized that it
makes the cost of authoring a unit test significantly lower. It's been
changing my opinion on how many unit tests/how much coverage we should
have. If the cost of a test is dramatically cheaper, would we want to
increase our coverage? Is it too extreme to say now we want 100%?

To that end, I just wanted to see what our coverage is today. This PR
adds a quick and dirty reporter, just to see how things look. I don't
think its useful to run it in CI or anything of that nature yet, until
we form a clearer opinion about what I mentioned in the previous
paragraph.

This implementation is based on
https://doc.rust-lang.org/rustc/instrument-coverage.html.

### Testing Instructions

Run the commands that have been added.

---------

Co-authored-by: Chris Olszewski <chris.olszewski@vercel.com>