feat: allow Watch Mode with `--root-turbo-json` (#10687)
### Description

Unblock Watch Mode from being used with `--root-turbo-json`. 

### Testing Instructions

I tested this manually by using `turbo watch lint
--root-turbo-json=./some-dir/turbo.json` in a `create-turbo` and
changing that file. The tasks would be picked up and re-ran the same as
if I omitted the `--root-turbo-json` and used a standard path.

I also tried it with a JSONC and it worked.