chore(turbo_json): lift `$TURBO_DEFAULT$` handling to turbo.json resolution (#10709)
### Description

This PR moves the handling of `$TURBO_DEFAULT$` to be the responsibility
of `turbo.json` resolution.

Previously it was handled by our SCM module which should only be focused
on SCM operations, not magical strings from Turborepo.

The changes to the daemon only affect the daemon input hashing logic
which is currently disabled.

Reviewing each commit independently is suggested.

### Testing Instructions

Added unit tests and existing tests.