feat(turbo_json): add `$TURBO_EXTENDS$` (#10763)
### Description

This PR adds a new DSL keyword for array task definitions that alters
the behavior when using `extends` in a package `turbo.json`. Usually a
field from a package task definition will completely override the root
field e.g.
Root: `{ "env": ["FOO"] }`
Package: `{ "env": ["BAR"] }`
Result: `{"env": ["BAR"] }`

If the package uses the new `$TURBO_EXTENDS$` the root field value will
be extended from instead of overridden. e.g.
Root: `{ "env": ["FOO"] }`
Package: `{ "env": ["$TURBO_EXTENDS$", "BAR"] }`
Result: `{"env": ["FOO", "BAR"] }`

This functionality allows for greater flexibility when specializing
package task definitions without needing to copy/paste the shared values
and remembering to update them if the root changed.

It is currently gated by `turboExtends` feature flag in the root
`turbo.json`.

Reviewing this each commit individually in this PR is *highly*
suggested. The actual feature add is all done in the final commit, the
rest are prefactors.

The primary prefactors are:
- Plumbing through future flags to be read at config time, but usable at
task resolution time. This is necessary since this is set at root and
needs to be known when reading task definitions from package
`turbo.json`s
- Adding a new intermediate step for task definition creation named
"processed". Removes a majority of the DSL concepts and validations that
are exclusive to a single field. This paves the way for future
extensions to the DSL.

Nerd note:
Technically, we could extend `$TURBO_EXTENDS$` to additional fields that
are monoids. Extension behavior is really:
```
let a = if extends {
  base
} else {
  mempty
}
a mappend b
```
The only interesting addition we could make would be around boolean
fields which has 2 different monoids: any and all. Which one to choose
would depend on the default value of the flag. e.g. `cache` defaults to
false so we would want an any (or) behavior. I am not suggesting we do
this, it is more confusing than it is worthwhile.

### Testing Instructions

Added a fair amount of unit tests for the whole task definition
resolution steps.