fix: adjust binary call for microfrontends proxy on Windows (#10962)
### Description

The microfrontends local development proxy fails to start on Windows
because it runs into: `%1 is not a valid Win32 application`. The issue
is that Turborepo runs the `.js` script and relies on the shebang to be
run correctly. That doesn't work on Windows. This PR changes the code to
run `node` which will work on Windows.

### Testing Instructions

In a microfrontends repository, such as
https://github.com/vercel-labs/microfrontends-nextjs-app-multi-zone, run
`pnpm turbo run dev` on Windows and ensure that it starts up.

---------

Co-authored-by: Anthony Shew <anthonyshew@gmail.com>