fix: Improve parity with real Bun lockfile format for `prune` (#11048)
### Description

Updating our Bun lockfile parser to achieve full parity with
`--frozen-lockfile`.

### Testing Instructions

I've added fixtures from #11007 and hand-tested each fixture with real
repositories with this newly updated logic. You'll also note that I've
removed some unit testing in favor of snapshot testing. Ultimately, we
care that the lockfile gets pruned and created correctly, which is what
the snapshots are testing for. Given Bun's relatively fast iteration
speed, the internals of how we compute the pruned lockfile don't help as
much for now. We can solidify with unit tests in the future if we need
to.