fix: Improve determinism for graceful shutdown (#11063)
### Description

My theory is we have a race condition when `child.stop()` is called and
child process exist quickly. We process the SIGINT in this
`tokio::select`. Depending on which branch of the select the branch goes
down, you could get the wrong behavior (KilledExternal instead of
Interrupted).

`biased;` gives us determinism for this situation.

### Testing Instructions

CI

<sub>CLOSES TURBO-4946</sub>