ci: Avoid flaking for test server mock (#11066)
### Description

The way these tests were written didn't necessarily wait for the server
to be ready, just _created_. A subtle distinction, but an important one
that was flaking in `test_http_cache`
([example](https://github.com/vercel/turborepo/actions/runs/19112365936/job/54612679302?pr=11053)).

I fixed it everywhere so it won't be a problem again.

### Testing Instructions

CI

<sub> CLOSES TURBO-4949</sub>