fix: Improve version resolution in Bun lockfiles (#11095)
### Description

We previously were not checking semver versions when building up the
lockfiles, so we had the potential to violate the semver ranges of the
original lockfile. Now, we do check them so that we ensure we're
preserving semver intent in the lockfile.

### Testing Instructions

I've tested this against the reproduction provided in #11007 and added
some more fixtures for CI.