ci: Fix flaky test (hopefully) (#11278)
### Description

This test has been flaky for awhile and I think I just figured out that
it has to do with the telemetry client not being set up yet when the
call to the implementation code gets made. When that happens, a log gets
sent into stdout, which the test is attempting to parse JSON from. When
that happens, stdout has a warning printed by the telemetry crate
instead of the JSON its expecting.

### Testing Instructions

CI over the next week or so.

<sub>CLOSES TURBO-5012</sub>