fix: Correctly validate workspace root as package (#11284)
### Description

#11258 shows a situation where we were interpreting the root
`turbo.json` as a Package Configuration when not using `extends`, and as
a root `turbo.json` when using `extends`. As described in the issue,
neither case works. This can happen when `.` is in the workspace
definition.

This is fixed by making sure our loaders/validators are properly
interpreting the `turbo.json`'s context correctly.

### Testing Instructions

Hand-tested that the repro works as expected and added some tests.